requirements.txt
faiss-cpu
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
pypdf
python-dotenv
sentence-transformers
streamlit
